improve logging
improve logging

add more logging output

check ~/Downloads for Logos icon file

improved logging

remove unused imports

specify MSI file name that is executed

+++ installer.py
@@ -33,7 +33,6 @@ from wine import initializeWineBottle
 from wine import install_msi
 from wine import installFonts
 from wine import installD3DCompiler
-from wine import run_logos
 from wine import wine_reg_install
 from wine import winetricks_install
 
@@ -414,6 +413,7 @@ def postInstall(app):
         logging.info(message)
 
         if not os.path.isfile(config.CONFIG_FILE): # config.CONFIG_FILE is set in main() function
+            logging.info(f"No config file at {config.CONFIG_FILE}")
             os.makedirs(os.path.join(HOME, ".config", "Logos_on_Linux"), exist_ok=True)
             if os.path.isdir(os.path.join(HOME, ".config", "Logos_on_Linux")):
                 write_config(config.CONFIG_FILE, config_keys)
@@ -421,7 +421,9 @@ def postInstall(app):
             else:
                 logos_warn(f"{HOME}/.config/Logos_on_Linux does not exist. Failed to create config file.")
         elif os.path.isfile(config.CONFIG_FILE):
+            logging.info(f"Config file exists at {config.CONFIG_FILE}.")
             # Compare existing config file contents with installer config.
+            logging.info(f"Comparing its contents with current config.")
             current_config_file_dict = config.get_config_file_dict(config.CONFIG_FILE)
             different = False
             for key in config_keys:
@@ -429,6 +431,7 @@ def postInstall(app):
                     different = True
                     break
             if different is True and logos_acknowledge_question(f"Update config file at {config.CONFIG_FILE}?", "The existing config file was not overwritten."):
+                logging.info(f"Updating config file.")
                 write_config(config.CONFIG_FILE, config_keys)
         else:
             # Script was run with a config file. Skip modifying the config.
@@ -440,7 +443,9 @@ def postInstall(app):
             launcher_exe = Path(f"{config.INSTALLDIR}/LogosLinuxLauncher")
             # FIXME: Confirm file copy and test desktop launcher.
             if launcher_exe.is_file():
+                logging.debug(f"Removing existing launcher binary.")
                 launcher_exe.unlink()
+            logging.info(f"Creating launcher binary by copying this installer binary to {launcher_exe}.")
             shutil.copy(sys.executable, launcher_exe)
             create_shortcut()
 
@@ -511,23 +516,33 @@ def create_shortcut():
         config.LOGOS_ICON_FILENAME = os.path.basename(config.LOGOS_ICON_URL)
 
     logos_icon_path = os.path.join(config.APPDIR, config.LOGOS_ICON_FILENAME)
+    logos_icon_downloaded = os.path.join(config.MYDOWNLOADS, config.LOGOS_ICON_FILENAME)
 
     if not os.path.isfile(logos_icon_path):
         os.makedirs(config.APPDIR, exist_ok=True)
-        raw_bytes = None
-        try:
-            with urllib.request.urlopen(config.LOGOS_ICON_URL) as f:
-                raw_bytes = f.read()
-        except urllib.error.URLError as e:
-            logos_error(e)
-        if raw_bytes is not None:
-            with open(logos_icon_path, 'wb') as f:
-                f.write(raw_bytes)
+        if not os.path.isfile(logos_icon_downloaded):
+            logging.info(f"Downloading icon from {config.LOGOS_ICON_URL}")
+            raw_bytes = None
+            try:
+                with urllib.request.urlopen(config.LOGOS_ICON_URL) as f:
+                    raw_bytes = f.read()
+            except urllib.error.URLError as e:
+                logos_error(e)
+            if raw_bytes is not None:
+                with open(logos_icon_path, 'wb') as f:
+                    f.write(raw_bytes)
+        else:
+            logging.info(f"Using icon file at {logos_icon_downloaded}")
+            shutil.copy(logos_icon_downloaded, logos_icon_path)
+    else:
+        logging.info(f"Icon found at {logos_icon_path}.")
 
     desktop_entry_path = os.path.expanduser(f"~/.local/share/applications/{config.FLPRODUCT}Bible.desktop")
     if os.path.exists(desktop_entry_path):
+        logging.info(f"Removing desktop launcher at {desktop_entry_path}.")
         os.remove(desktop_entry_path)
 
+    logging.info(f"Creating desktop launcher at {desktop_entry_path}.")
     with open(desktop_entry_path, 'w') as desktop_file:
         desktop_file.write(f"""[Desktop Entry]
 Name={config.FLPRODUCT}Bible
